# speed up `tunable.model_spec()`